+++ src/RcppExports.cpp
@@ -66,6 +66,27 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// force_bundle_iter
+List force_bundle_iter(NumericMatrix edges_xy, List elist, double K, int C, int P, int P_rate, double S, int I, double I_rate, double compatibility_threshold, double eps);
+RcppExport SEXP _ggraph_force_bundle_iter(SEXP edges_xySEXP, SEXP elistSEXP, SEXP KSEXP, SEXP CSEXP, SEXP PSEXP, SEXP P_rateSEXP, SEXP SSEXP, SEXP ISEXP, SEXP I_rateSEXP, SEXP compatibility_thresholdSEXP, SEXP epsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type edges_xy(edges_xySEXP);
+    Rcpp::traits::input_parameter< List >::type elist(elistSEXP);
+    Rcpp::traits::input_parameter< double >::type K(KSEXP);
+    Rcpp::traits::input_parameter< int >::type C(CSEXP);
+    Rcpp::traits::input_parameter< int >::type P(PSEXP);
+    Rcpp::traits::input_parameter< int >::type P_rate(P_rateSEXP);
+    Rcpp::traits::input_parameter< double >::type S(SSEXP);
+    Rcpp::traits::input_parameter< int >::type I(ISEXP);
+    Rcpp::traits::input_parameter< double >::type I_rate(I_rateSEXP);
+    Rcpp::traits::input_parameter< double >::type compatibility_threshold(compatibility_thresholdSEXP);
+    Rcpp::traits::input_parameter< double >::type eps(epsSEXP);
+    rcpp_result_gen = Rcpp::wrap(force_bundle_iter(edges_xy, elist, K, C, P, P_rate, S, I, I_rate, compatibility_threshold, eps));
+    return rcpp_result_gen;
+END_RCPP
+}
 // hTree
 NumericMatrix hTree(IntegerVector parent, IntegerVector order);
 RcppExport SEXP _ggraph_hTree(SEXP parentSEXP, SEXP orderSEXP) {
@@ -161,6 +182,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_ggraph_pack", (DL_FUNC) &_ggraph_pack, 1},
     {"_ggraph_circlePackLayout", (DL_FUNC) &_ggraph_circlePackLayout, 2},
     {"_ggraph_dendrogram_spread", (DL_FUNC) &_ggraph_dendrogram_spread, 7},
+    {"_ggraph_force_bundle_iter", (DL_FUNC) &_ggraph_force_bundle_iter, 11},
     {"_ggraph_hTree", (DL_FUNC) &_ggraph_hTree, 2},
     {"_ggraph_partitionTree", (DL_FUNC) &_ggraph_partitionTree, 4},
     {"_ggraph_cut_lines", (DL_FUNC) &_ggraph_cut_lines, 9},