+++ NEWS.md
@@ -27,6 +27,9 @@
 * Updated layout functions from the graphlayouts package to support grouped and
   constrained versions (centrality, focus, and stress layouts)
 * Added H Tree layout for binary trees (#58)
+* Added `geom_edge_bundle_force()`, `geom_edge_bundle_path()`, and 
+  `geom_edge_bundle_minimal()` (+ variants) to provide support for edge bundling 
+  (#267)
 
 # ggraph 2.1.0